quota: remove quota utils
