+++ test/extended/imageapis/helper.go
@@ -1,9 +1,9 @@
-package util
+package imageapis
 
 import (
 	"strings"
 
-	apierrs "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/errors"
 )
 
 // errMessageString is a part of error message copied from quotaAdmission.Admit() method in
@@ -15,7 +15,7 @@ const errLimitsMessageString = `exceeds the maximum limit`
 // IsErrorQuotaExceeded returns true if the given error stands for a denied request caused by detected quota
 // abuse.
 func IsErrorQuotaExceeded(err error) bool {
-	if isForbidden := apierrs.IsForbidden(err); isForbidden || apierrs.IsInvalid(err) {
+	if isForbidden := errors.IsForbidden(err); isForbidden || errors.IsInvalid(err) {
 		lowered := strings.ToLower(err.Error())
 		// the limit error message can be accompanied only by Invalid reason
 		if strings.Contains(lowered, errLimitsMessageString) {
@@ -23,18 +23,6 @@ func IsErrorQuotaExceeded(err error) bool {
 		}
 		// the quota error message can be accompanied only by Forbidden reason
 		if isForbidden && (strings.Contains(lowered, errQuotaMessageString) || strings.Contains(lowered, errQuotaUnknownMessageString)) {
-			return true
-		}
-	}
-	return false
-}
-
-// IsErrorLimitExceeded returns true if the given error is a limit error.
-func IsErrorLimitExceeded(err error) bool {
-	if isForbidden := apierrs.IsForbidden(err); isForbidden || apierrs.IsInvalid(err) {
-		lowered := strings.ToLower(err.Error())
-		// the limit error message can be accompanied only by Invalid reason
-		if strings.Contains(lowered, errLimitsMessageString) {
 			return true
 		}
 	}
